requirements.txt
langchain
langchain-openai
python-dotenv
rich
flask
flask-cors
PyPDF2
python-docx
faiss-cpu
langchain_community
pydantic<2.0
scikit-learn
